moved middlewares to server pkg and added server key for context key
